Review comments.

+++ programs/mango-v4/src/logs.rs
@@ -213,6 +213,6 @@ pub struct DeactivateTokenPositionLog {
     pub mango_group: Pubkey,
     pub mango_account: Pubkey,
     pub token_index: u16,
-    pub cumulative_deposit_interest: i64,
-    pub cumulative_borrow_interest: i64,
+    pub cumulative_deposit_interest: f32,
+    pub cumulative_borrow_interest: f32,
 }
